pyproject.toml
[project]
name = "gos-sanity"
version = "0.2.0"
description = "80-rollout sanity check: does SkillsBench reward respond to bundle composition?"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.2",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "matplotlib>=3.8",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
